Add more tests for the meet operator

+++ tests/regression-tests/expected-outputs/sparc/fbranch2/fbranch2/fbranch2.c
@@ -3,8 +3,8 @@ int main(int argc, char *argv[]);
 /** address: 0x00010694 */
 int main(int argc, char *argv[])
 {
-    union { float; int; } f8; 		// r40
-    union { float; int; } local0; 		// m[o6 - 24]
+    union { int; float; } f8; 		// r40
+    union { int; float; } local0; 		// m[o6 - 24]
     union { __size32; double; } local1; 		// m[o6 - 12]
 
     f8 = *0x10940;
